Amethyst installed and configured
