+++ home/config/.config/hypr/hyprland.conf
@@ -123,7 +123,6 @@ windowrule = float,^(alacritty)$
 windowrule = float,^(pavucontrol)$
 windowrule = float,^(blueman-manager)$
 windowrulev2 = stayfocused, class:^(pinentry-) # fix pinentry losing focus
-
 $mainMod = SUPER_CTRL_ALT #mapped to caps w/ kmonad
 
 # Apps
