+++ home/config/.hammerspoon/apps.lua
@@ -1,6 +1,3 @@
-local super = {"ctrl", "alt", "cmd"}
-local superShift = {"ctrl", "alt", "cmd", "lshift"}
-
 --- start quick open applications 
 function open_app(name)
     return function()
