+++ home/config/.hammerspoon/commands.lua
@@ -1,6 +1,3 @@
-local super = {"ctrl", "alt", "cmd"}
-local superShift = {"ctrl", "alt", "cmd", "lshift"}
-
 
 -- Commands
 hs.hotkey.bind(super, "q", nil, function() hs.eventtap.keyStroke({"cmd"}, "q") end)
