+++ home/config/.hammerspoon/readme.md
@@ -1,0 +1,4 @@
+# Yabai vs Amethyst
+
+When I was building this configuration, there were issues with Yabai and 14.4 so I configured amethyst via brew instead. 
+I'm leaving the yabai.lua in case I go back.
