+++ system/darwin/hosts/mwdavis-workm1/default.nix
@@ -5,7 +5,7 @@
   nyx = {
     modules = {
       user.home = ./home.nix;
-      system.yabai.enable = true;
+      system.yabai.enable = false;
     };
 
     secrets = {
